Renames PeriodMeterDataCOntainer -> 0PeriodMeterDataContainer and MeterData -> PeriodMeterData
